Avoid creation of dangling symlinks
If rsync fails, possibly the new directory does not exist, so we should
not create the symlink. Because if we create it, the next backup would
not be incremental, because the --link-dest option does not point to an
existing directory.

This patch fixes this by checking the return code of rsync and only
creates the symlink if rsync exits successful or with error 24 ("Partial
transfer due to vanished source files") which is somewhat expected.

Signed-off-by: Rodrigo Campos <rodrigo@sdfg.com.ar>